requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0
fastapi==0.103.1
uvicorn==0.23.2
aiohttp==3.8.5
python-multipart
google-generativeai==0.3.1